Water simulation failed project

+++ GLRenderer/GLRenderer/GLRendererInterface2D.h
@@ -9,7 +9,7 @@ public:
 	explicit C_GLRendererInterface2D(const C_GLRendererInterface2D& other) = delete;
 	explicit C_GLRendererInterface2D(C_GLRendererInterface2D&& other) noexcept
 	{
-		m_VAOid = other.m_VAOid;
+		m_VAOid			  = other.m_VAOid;
 		m_PlaneVertHandle = other.m_PlaneVertHandle;
 	}
 	~C_GLRendererInterface2D() override;
@@ -20,6 +20,6 @@ private:
 	GLuint							   m_VAOid; // one generic buffer for all calls to ignore VAOs
 	Renderer::Handle<Renderer::Buffer> m_PlaneVertHandle;
 	Renderer::Handle<Renderer::Buffer> m_UVVertHandle;
-	glm::vec2						   m_RenderTargetSize;
+	glm::uvec2						   m_RenderTargetSize;
 };
 } // namespace GLEngine::GLRenderer
